requirements.txt
PyQt6
numpy
pyqtdarktheme
scikit-image
Pillow
dlib
qdarktheme